requirements.txt
fastapi==0.115.6
pydantic==2.10.3
psycopg[binary]==3.2.10
PyYAML==6.0.2
Jinja2==3.1.6
httpx==0.28.1
uvicorn==0.32.1
